[BugFix][Text] Fix vertical offset issue in multi-line text rendering.
After setting the line-height, we will center the inline text, but currently the maximum ascent and minimum descent used are for all text fonts, resulting in abnormal offsets in some scenarios, such as nested views.

Calculate the maximum ascent and minimum descent using the font of the line text.

issue:m-5839873515