[Optimize] lepus::Dictionary value get interface
Currently lepus::Dictionary provides GetValue, GetProperty and At methods to get value. This commit will unify the getter methods to GetValue and GetValueOrXXX design which is more readable and self-explainable.

Introduce ValueWrapper and change the return value from `const Value&` to `ValueWrapper` to reduce the chance that inner Value pointer address is fetched and stored elsewhere. This is the preparation for later optimizations.

issue: m-6481490102