[Optimize] ElementTemplate supports encoding and decoding element config.
Described as the title. Handle element config as a built-in attribute, supporting the framework to pass the config during the generation of the element template binary. The Lynx encoder will record it in the built-in attribute.

During the execution of the decode logic, the config will also be stored in the built-in attribute. When building the element, if the built-in attribute has a config, it will be assigned to the element's config_.

issue:m-4872381525
AutoSubmit:True