[Optimize][DevTool] Align the path of switch page in Android and iOS.
In this change, I align the paths of the page files in the two systems so that in the future more work could be done in front-end without worrying about the difference.
To achieve that, I remove the extra folder in Android.
Yet in order not to break applications who read that switch page file, I duplicate a file as a deprecation approach for potential adaptions.