[Feature] Add the tools for reporting build system time consumption.
Summary:

By analyzing the performance reports automatically generated by the build tools on each platform, we obtain data such as the top 20 time-consuming subtasks, the total time consumption, the total number of C++ files involved in the build, and the top 8 C++ files with the highest number of repeated compilations. The obtained build information is then output to a JSON file in the trace event format. Finally, it is reported to the company's database through the trace report of CQ CDS. Subsequently, the analysis data set can be viewed on the data visualization platform. In addition, the performance reports related to the builds on each platform are packaged and uploaded to the object storage platform for subsequent download and detailed data review.

Key modifications:

1. Add a parsing tool for Android build time consumption files. It is used to parse the profile.html performance report automatically generated by Gradle and generate the information and resources required for reporting.
2. Add a parsing tool for iOS build time consumption files. It is used to parse the xcactivitylog performance report automatically generated by xcodebuild and generate the information and resources required for reporting.
3. Add a parsing tool for GN build time consumption files. It is used to parse build log files such as ninja_log and generate the information and resources required for reporting.
4. Optimize the path for generating the lynx-project-info file in package.gradle.
5. Add logic to output stage time consumption in bundle_install.sh and gn.py.
6. Add tool test scripts.

AutoSubmit:true