[BugFix] Fix the wrong scroll-state transition of scrollable
When pointer cursor is on a scrollable which is the child of another scrollable, we rotate the mouse
wheel. If child can't scroll in the direction of wheel event, we try to scroll parent. However, in
this case, the scroll status of parent isn't set properly. This MR is fixing this bug.

issue: m-6796016008